기본 상황에서의 시그널 완성
