+++ Makefile
@@ -6,7 +6,7 @@
 #    By: sehhong <sehhong@student.42seoul.kr>       +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2022/03/12 14:49:00 by jiskim            #+#    #+#              #
-#    Updated: 2022/04/08 01:12:00 by sehhong          ###   ########.fr        #
+#    Updated: 2022/04/08 15:23:57 by sehhong          ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
@@ -74,9 +74,7 @@ SRCS_AST	= $(addprefix $(SRCS_AST_DIR),\
 				read_ast.c \
 				)
 SRCS_SIG	= $(addprefix $(SRCS_SIG_DIR), \
-				sigint.c \
-				sigquit.c \
-				tc_cursor.c \
+				signal_handler.c \
 				tc_echoctl.c \
 				)
 
@@ -134,7 +132,7 @@ all: $(NAME)
 	@printf $(UP)$(UP)
 
 $(NAME) : $(OBJS) $(LIB)
-	@$(CC) $(CFLAGS) $(RDLN_LFLAGS) -lcurses $(LIB) $^ -o $@
+	@$(CC) $(CFLAGS) $(RDLN_LFLAGS) $(LIB) $^ -o $@
 	@printf $(CUT)$(DOWN)$(CUT)
 	@echo $(BOLD)$(L_PURPLE) 🍓 strawberry shell is $(L_RED)ready!!$(RESET)
