+++ includes/tc_signal.h
@@ -6,7 +6,7 @@
 /*   By: sehhong <sehhong@student.42seoul.kr>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/04/07 22:07:41 by sehhong           #+#    #+#             */
-/*   Updated: 2022/04/08 03:28:21 by sehhong          ###   ########.fr       */
+/*   Updated: 2022/04/08 15:15:46 by sehhong          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,14 +15,7 @@
 
 void	sigint_handler_p(int signum);
 void	sigint_handler_c(int signum);
-
-// void	sigquit_handler(int signum);
-
-/* 커서 제어 관련 함수 -> 안쓸것같음...ㅠㅠ
-void	initial_config(void);
-void	get_cursor_pos(int *x, int *y);
-void	action_for_eof(void);
-*/
+void	sigquit_handler(int signum);
 
 void	off_echoctl(void);
 void	on_echoctl(void);
