+++ srcs/ast/read_ast.c
@@ -6,7 +6,7 @@
 /*   By: sehhong <sehhong@student.42seoul.kr>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/03/28 14:45:00 by sehhong           #+#    #+#             */
-/*   Updated: 2022/04/08 03:02:43 by sehhong          ###   ########.fr       */
+/*   Updated: 2022/04/08 14:37:46 by sehhong          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -61,8 +61,6 @@ void	read_ast(t_ast *tree)
 	{	
 		on_echoctl();
 		signal(SIGINT, sigint_handler_c);
-		// signal(SIGQUIT, sigquit_handler);
-		signal(SIGQUIT, SIG_DFL);
 		wait_children();
 		free_cmd_list();
 		return ;
