--- srcs/signal/sigint.c
@@ -1,32 +1,0 @@
-/* ************************************************************************** */
-/*                                                                            */
-/*                                                        :::      ::::::::   */
-/*   sigint.c                                           :+:      :+:    :+:   */
-/*                                                    +:+ +:+         +:+     */
-/*   By: sehhong <sehhong@student.42seoul.kr>       +#+  +:+       +#+        */
-/*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2022/04/07 22:04:17 by sehhong           #+#    #+#             */
-/*   Updated: 2022/04/07 22:52:46 by sehhong          ###   ########.fr       */
-/*                                                                            */
-/* ************************************************************************** */
-
-#include "minishell.h"
-
-void	sigint_handler_p(int signum)
-{
-	if (signum != SIGINT)
-		return ;
-	g_box->exit_code = EXIT_FAILURE;
-	write(STDOUT_FILENO, "\n", 1);
-	rl_on_new_line();
-	rl_replace_line("", 1);
-	rl_redisplay();
-}
-
-void	sigint_handler_c(int signum)
-{
-	if (signum != SIGINT)
-		return ;
-	write(STDOUT_FILENO, "\n", 1);
-	return ;
-}
